Add files via upload
For this project we see a switched LAN/VLAN with two networks representing company departments. Dynamically assigned IP addresses through our DHCP server and the router configured as a relay agent to support DHCP to assign and identify host IP addresses. Router is configured to run RIP version 2 so that network traffic is routed effectively. Here we only have a LAN so that traffic isn't traversing a WAN and remains local. The router is configured with basic security controls at the console and virtual lines to require valid authentication from admin. Our two departments are on separate networks and VLANs so we have a dot1q trunk link configured between our switch and router should any inter-VLAN routed need to be supported.