Add files via upload
Here is a simple Bash script for an address book scripted in Nano. Has a menu that prompts the user to Create, search or view a contact in the address book. The required permissions were provided through the chmod command for this script to be executed in the terminal. I like Nano for it's user-friendly functions and feels a lot like Microsoft word for Linux scripts. As I progress and take a deeper dive into Linux I know Vi is another preferential text editor thats popular amongst users.