Add files via upload
Simple Wireshark SSH (Secure Shell) traffic analysis via the Linux terminal. Here the GUI was accessed in Lunix and failed SSH attempts from the personal home laptop to the Linux VM are recorded. The three way handshake SYN-SYN/ACK-SYN isn't completed and the reason for that is invalid password attempts with the Eliptic Curve Cryptography. Wireshark provides valuable analysis of source and dest IP and detects and records the failed attempts with valuable data that presents a clear picture of network traffic, both permitted and denied. We see source and destination ports with the tagged TCP protocol. The SSH filter was used to only source SSH traffic on the machine.