Add files via upload
Interesting project here where a file was uploaded that recorded an unauthorized Telnet session on a company device. This project required Linux and Wireshark utilities to conduct analysis on the traffic and determine how this access occurred. A few screenshots show ARP requests to the target device and eventually Telent request through opn port 23 on this network device. Analysis of the traffic provides a range of information including protocols used, source and destination IP, confirms unauthorized access was gained and a Linux commands were entered in an attempt to access password files by cat'ing the etc/passwd and etc/shadow files where user account passwords are stored in Linux environments.